+++ src/ruleset/era.rs
@@ -4,15 +4,15 @@ use serde::{Deserialize, Serialize};
 #[serde(rename_all = "camelCase")]
 pub struct EraInfo {
     name: String,
-    research_agreement_cost: i16,
-    starting_settler_count: i8,
-    starting_worker_count: i8,
-    starting_military_unit_count: i8,
+    research_agreement_cost: i32,
+    starting_settler_count: i32,
+    starting_worker_count: i32,
+    starting_military_unit_count: i32,
     starting_military_unit: String,
-    settler_population: i8,
-    base_unit_buy_cost: i16,
-    embark_defense: i8,
-    start_percent: i8,
+    settler_population: i32,
+    base_unit_buy_cost: i32,
+    embark_defense: i32,
+    start_percent: i32,
     city_sound: String,
     #[serde(rename = "iconRGB")]
     icon_rgb: [u8; 3],
